+++ src/app/(private)/(home)/stores/transactionList.ts
@@ -5,7 +5,7 @@ import { Transaction } from "../types/transaction";
 ////////////////////////////////////////////////////////////////////////////////////////////////
 // アトムとセレクター
 ////////////////////////////////////////////////////////////////////////////////////////////////
-export const transactionListState = atom({
+export const transactionListState = atom<Transaction[]>({
     key: "transactionListState",
     default: [], // 初期値は空のリスト
 });
